Create utils/models.py
