+++ wagtail/core/models/__init__.py
@@ -1,45 +1,3 @@
-from wagtail.models import (  # noqa
-    COMMENTS_RELATION_NAME,
-    PAGE_MODEL_CLASSES,
-    PAGE_PERMISSION_TYPE_CHOICES,
-    PAGE_PERMISSION_TYPES,
-    PAGE_TEMPLATE_VAR,
-    AbstractPage,
-    BasePageManager,
-    Comment,
-    CommentReply,
-    GroupApprovalTask,
-    GroupPagePermission,
-    Orderable,
-    Page,
-    PageBase,
-    PageLogEntry,
-    PageLogEntryManager,
-    PageLogEntryQuerySet,
-    PageManager,
-    PagePermissionTester,
-    PageRevision,
-    PageSubscription,
-    PageViewRestriction,
-    SubmittedRevisionsManager,
-    Task,
-    TaskManager,
-    TaskState,
-    TaskStateManager,
-    UserPagePermissionsProxy,
-    Workflow,
-    WorkflowManager,
-    WorkflowPage,
-    WorkflowState,
-    WorkflowStateManager,
-    WorkflowTask,
-    get_default_page_content_type,
-    get_page_models,
-    get_streamfield_names,
-    logger,
-    reassign_root_page_locale_on_delete,
-)
-
 from .audit_log import (  # noqa
     BaseLogEntry,
     BaseLogEntryManager,
