--- wagtail/core/utils.py
@@ -1,19 +1,0 @@
-from wagtail.utils.coreutils import (  # noqa
-    SCRIPT_RE,
-    SLUGIFY_RE,
-    WAGTAIL_APPEND_SLASH,
-    InvokeViaAttributeShortcut,
-    accepts_kwarg,
-    camelcase_to_underscore,
-    cautious_slugify,
-    escape_script,
-    find_available_slug,
-    get_content_languages,
-    get_content_type_label,
-    get_model_string,
-    get_supported_content_language_variant,
-    multigetattr,
-    resolve_model_string,
-    safe_snake_case,
-    string_to_ascii,
-)
