+++ wagtail/tests/tests.py
@@ -6,10 +6,10 @@ from django.test.utils import override_settings
 from django.urls.exceptions import NoReverseMatch
 from django.utils.safestring import SafeString
 
+import wagtail.utils.models
 from wagtail.models import Locale, Page, Site, SiteRootPath
 from wagtail.templatetags.wagtailcore_tags import richtext, slugurl
 from wagtail.test.testapp.models import SimplePage
-from wagtail.utils.coreutils import resolve_model_string
 
 
 class TestPageUrlTags(TestCase):
@@ -355,47 +355,62 @@ class TestSiteRootPathsCache(TestCase):
 
 class TestResolveModelString(TestCase):
     def test_resolve_from_string(self):
-        model = resolve_model_string("wagtailcore.Page")
+        model = wagtail.utils.models.resolve_model_string("wagtailcore.Page")
 
         self.assertEqual(model, Page)
 
     def test_resolve_from_string_with_default_app(self):
-        model = resolve_model_string("Page", default_app="wagtailcore")
+        model = wagtail.utils.models.resolve_model_string(
+            "Page", default_app="wagtailcore"
+        )
 
         self.assertEqual(model, Page)
 
     def test_resolve_from_string_with_different_default_app(self):
-        model = resolve_model_string("wagtailcore.Page", default_app="wagtailadmin")
+        model = wagtail.utils.models.resolve_model_string(
+            "wagtailcore.Page", default_app="wagtailadmin"
+        )
 
         self.assertEqual(model, Page)
 
     def test_resolve_from_class(self):
-        model = resolve_model_string(Page)
+        model = wagtail.utils.models.resolve_model_string(Page)
 
         self.assertEqual(model, Page)
 
     def test_resolve_from_string_invalid(self):
-        self.assertRaises(ValueError, resolve_model_string, "wagtail.core.Page")
+        self.assertRaises(
+            ValueError, wagtail.utils.models.resolve_model_string, "wagtail.core.Page"
+        )
 
     def test_resolve_from_string_with_incorrect_default_app(self):
         self.assertRaises(
-            LookupError, resolve_model_string, "Page", default_app="wagtailadmin"
+            LookupError,
+            wagtail.utils.models.resolve_model_string,
+            "Page",
+            default_app="wagtailadmin",
         )
 
     def test_resolve_from_string_with_unknown_model_string(self):
-        self.assertRaises(LookupError, resolve_model_string, "wagtailadmin.Page")
+        self.assertRaises(
+            LookupError, wagtail.utils.models.resolve_model_string, "wagtailadmin.Page"
+        )
 
     def test_resolve_from_string_with_no_default_app(self):
-        self.assertRaises(ValueError, resolve_model_string, "Page")
+        self.assertRaises(ValueError, wagtail.utils.models.resolve_model_string, "Page")
 
     def test_resolve_from_class_that_isnt_a_model(self):
-        self.assertRaises(ValueError, resolve_model_string, object)
+        self.assertRaises(ValueError, wagtail.utils.models.resolve_model_string, object)
 
     def test_resolve_from_bad_type(self):
-        self.assertRaises(ValueError, resolve_model_string, resolve_model_string)
+        self.assertRaises(
+            ValueError,
+            wagtail.utils.models.resolve_model_string,
+            wagtail.utils.models.resolve_model_string,
+        )
 
     def test_resolve_from_none(self):
-        self.assertRaises(ValueError, resolve_model_string, None)
+        self.assertRaises(ValueError, wagtail.utils.models.resolve_model_string, None)
 
 
 class TestRichtextTag(TestCase):
